strip newlines from twitter embed
Not sure if this is the best way of handling it, but it worked for me. Background on this is that Jekyll was choking on the newline when a tweet was included as part of a list item, causing the list to close early and leaving a dangling `</div>`

The Twitter embed comes back with `\n` between the blockquote and the script. For whatever reason Jekyll hated this. ¯\_(ツ)_/¯